+++ test/curation_test.exs
@@ -33,7 +33,7 @@ defmodule CurationTest do
     %{schema: schema, conn: conn, map_conn: map_conn}
   end
 
-  @tag ["27.1": true, "27.0": true, "26.0": true]
+  @tag ["28.0": true, "27.1": true, "27.0": true, "26.0": true]
   test "success: get override", %{schema: schema, conn: conn, map_conn: map_conn} do
     body =
       %{
@@ -85,7 +85,7 @@ defmodule CurationTest do
     assert {:ok, _} = ExTypesense.get_override(map_conn, House, name, [])
   end
 
-  @tag ["27.1": true, "27.0": true, "26.0": true]
+  @tag ["28.0": true, "27.1": true, "27.0": true, "26.0": true]
   test "success: list all overrides", %{schema: schema, conn: conn, map_conn: map_conn} do
     assert {:ok, %SearchOverridesResponse{}} = ExTypesense.list_overrides(schema.name)
     assert {:ok, _} = ExTypesense.list_overrides(schema.name, [])
@@ -102,7 +102,7 @@ defmodule CurationTest do
     assert {:ok, _} = ExTypesense.list_overrides(map_conn, House, [])
   end
 
-  @tag ["27.1": true, "27.0": true, "26.0": true]
+  @tag ["28.0": true, "27.1": true, "27.0": true, "26.0": true]
   test "success: delete override", %{schema: schema, conn: conn, map_conn: map_conn} do
     body =
       %{