+++ test/key_test.exs
@@ -28,7 +28,7 @@ defmodule KeyTest do
     %{api_key_schema: api_key_schema, conn: conn, map_conn: map_conn}
   end
 
-  @tag ["27.1": true, "27.0": true, "26.0": true]
+  @tag ["28.0": true, "27.1": true, "27.0": true, "26.0": true]
   test "success: get keys", %{api_key_schema: api_key_schema, conn: conn, map_conn: map_conn} do
     assert {:ok, api_key} = ExTypesense.create_key(api_key_schema)
     assert {:ok, _} = ExTypesense.create_key(api_key_schema, [])
@@ -47,7 +47,7 @@ defmodule KeyTest do
     assert {:ok, %ApiKey{id: ^key_id}} = ExTypesense.get_key(map_conn, key_id, [])
   end
 
-  @tag ["27.1": true, "27.0": true, "26.0": true]
+  @tag ["28.0": true, "27.1": true, "27.0": true, "26.0": true]
   test "success: delete keys", %{api_key_schema: api_key_schema, conn: conn, map_conn: map_conn} do
     assert {:ok, api_key} = ExTypesense.create_key(api_key_schema)
     key_id = api_key.id
@@ -60,7 +60,7 @@ defmodule KeyTest do
     assert {:error, _} = ExTypesense.delete_key(map_conn, key_id, [])
   end
 
-  @tag ["27.1": true, "27.0": true, "26.0": true]
+  @tag ["28.0": true, "27.1": true, "27.0": true, "26.0": true]
   test "success: list all keys", %{conn: conn, map_conn: map_conn} do
     assert {:ok, %ApiKeysResponse{}} = ExTypesense.list_keys()
     assert {:ok, _} = ExTypesense.list_keys([])