+++ test/stopwords_test.exs
@@ -3,7 +3,6 @@ defmodule StopwordsTest do
 
   alias OpenApiTypesense.ApiResponse
   alias OpenApiTypesense.Connection
-  alias OpenApiTypesense.Stopwords
   alias OpenApiTypesense.StopwordsSetRetrieveSchema
   alias OpenApiTypesense.StopwordsSetSchema
   alias OpenApiTypesense.StopwordsSetsRetrieveAllSchema
@@ -15,7 +14,7 @@ defmodule StopwordsTest do
     %{conn: conn, map_conn: map_conn}
   end
 
-  @tag ["27.1": true, "27.0": true, "26.0": true]
+  @tag ["28.0": true, "27.1": true, "27.0": true, "26.0": true]
   test "success: get stopword", %{conn: conn, map_conn: map_conn} do
     stop_id = "stopword_set_countries"
 
@@ -43,7 +42,7 @@ defmodule StopwordsTest do
     assert {:ok, _} = ExTypesense.get_stopword(map_conn, stop_id, [])
   end
 
-  @tag ["27.1": true, "27.0": true, "26.0": true]
+  @tag ["28.0": true, "27.1": true, "27.0": true, "26.0": true]
   test "success: list all stopwords", %{conn: conn, map_conn: map_conn} do
     assert {:ok, %StopwordsSetsRetrieveAllSchema{}} = ExTypesense.list_stopwords()
     assert {:ok, _} = ExTypesense.list_stopwords([])
@@ -53,7 +52,7 @@ defmodule StopwordsTest do
     assert {:ok, _} = ExTypesense.list_stopwords(map_conn, [])
   end
 
-  @tag ["27.1": true, "27.0": true, "26.0": true]
+  @tag ["28.0": true, "27.1": true, "27.0": true, "26.0": true]
   test "success: delete stopword", %{conn: conn, map_conn: map_conn} do
     stop_id = "stopword_set_countries"
 
@@ -64,7 +63,7 @@ defmodule StopwordsTest do
       }
 
     assert {:ok, %StopwordsSetSchema{id: ^stop_id}} = ExTypesense.upsert_stopword(stop_id, body)
-    assert {:ok, %Stopwords{id: ^stop_id}} = ExTypesense.delete_stopword(stop_id)
+    assert {:ok, %OpenApiTypesense.Stopwords{id: ^stop_id}} = ExTypesense.delete_stopword(stop_id)
 
     assert {:error, %ApiResponse{message: "Stopword `stopword_set_countries` not found."}} =
              ExTypesense.delete_stopword(stop_id, [])